+++ project/GenMustMatchersTests.scala
@@ -56,41 +56,43 @@ trait GenMustMatchersTestsBase {
     matchersDir.mkdirs()
 
     def transformFile(shouldFile: File, mustFile: File) {
-      val writer = new BufferedWriter(new FileWriter(mustFile))
-      try {
-        val shouldLines = Source.fromFile(shouldFile).getLines().toList // for 2.8
-        var skipMode = false
-        for (shouldLine <- shouldLines) {
-          val mustLine: String =
-            if (scalaJS) {
-              if (shouldLine.trim == "// SKIP-SCALATESTJS-START") {
-                skipMode = true
-                ""
-              }
-              else if (shouldLine.trim == "// SKIP-SCALATESTJS-END") {
-                skipMode = false
-                ""
-              }
-              else if (!skipMode) {
-                if (shouldLine.trim.startsWith("//SCALATESTJS-ONLY "))
-                  translateShouldToMustInTests(shouldLine.substring(shouldLine.indexOf("//SCALATESTJS-ONLY ") + 19))
+      if (!mustFile.exists || shouldFile.lastModified > mustFile.lastModified) {
+        val writer = new BufferedWriter(new FileWriter(mustFile))
+        try {
+          val shouldLines = Source.fromFile(shouldFile).getLines().toList // for 2.8
+          var skipMode = false
+          for (shouldLine <- shouldLines) {
+            val mustLine: String =
+              if (scalaJS) {
+                if (shouldLine.trim == "// SKIP-SCALATESTJS-START") {
+                  skipMode = true
+                  ""
+                }
+                else if (shouldLine.trim == "// SKIP-SCALATESTJS-END") {
+                  skipMode = false
+                  ""
+                }
+                else if (!skipMode) {
+                  if (shouldLine.trim.startsWith("//SCALATESTJS-ONLY "))
+                    translateShouldToMustInTests(shouldLine.substring(shouldLine.indexOf("//SCALATESTJS-ONLY ") + 19))
+                  else
+                    translateShouldToMustInTests(shouldLine)
+                }
                 else
-                  translateShouldToMustInTests(shouldLine)
+                  ""
               }
               else
-                ""
-            }
-            else
-              translateShouldToMustInTests(shouldLine)
+                translateShouldToMustInTests(shouldLine)
 
-          writer.write(mustLine.toString)
-          writer.newLine() // add for 2.8
+            writer.write(mustLine.toString)
+            writer.newLine() // add for 2.8
+          }
         }
-      }
-      finally {
-        writer.flush()
-        writer.close()
-        println("Generated " + mustFile.getAbsolutePath)
+        finally {
+          writer.flush()
+          writer.close()
+          println("Generated " + mustFile.getAbsolutePath)
+        }
       }
     }
